fix recurringInvoices modal
