+++ frontend/src/pages/(app)/invoices/_components/invoice-upsert.tsx
@@ -1,24 +1,20 @@
 import type { Client, Invoice, PaymentMethod, Quote } from "@/types"
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog"
-import { DndContext, MouseSensor, TouchSensor, closestCenter, useSensor, useSensors } from "@dnd-kit/core"
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form"
-import { GripVertical, Plus, Trash2 } from "lucide-react"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
-import { SortableContext, arrayMove, useSortable, verticalListSortingStrategy } from "@dnd-kit/sortable"
 import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { useEffect, useState } from "react"
-import { useFieldArray, useForm } from "react-hook-form"
+import { useForm } from "react-hook-form"
 import { useGet, usePatch, usePost } from "@/hooks/use-fetch"
 
 import { BetterInput } from "@/components/better-input"
 import { Button } from "@/components/ui/button"
-import { CSS } from "@dnd-kit/utilities"
+import { ClientSelectField } from "./client-select-field"
 import { ClientUpsert } from "../../clients/_components/client-upsert"
 import CurrencySelect from "@/components/currency-select"
 import { DatePicker } from "@/components/date-picker"
-import { Input } from "@/components/ui/input"
+import { InvoiceLineItemsEditor } from "./invoice-line-items-editor"
 import { PaymentMethodType } from "@/types"
-import type React from "react"
 import SearchSelect from "@/components/search-input"
 import { Separator } from "@/components/ui/separator"
 import { Switch } from "@/components/ui/switch"
@@ -34,6 +30,41 @@ interface InvoiceUpsertDialogProps {
 }
 
 type CreationMode = "invoice" | "recurring"
+
+function createItemSchema(t: (key: string) => string, translationPrefix: "invoices" | "recurringInvoices", typeSchema: z.ZodTypeAny) {
+    return z.object({
+        id: z.string().optional(),
+        description: z
+            .string()
+            .min(1, t(`${translationPrefix}.upsert.form.items.description.errors.required`))
+            .refine((val) => val !== "", {
+                message: t(`${translationPrefix}.upsert.form.items.description.errors.required`),
+            }),
+        type: typeSchema,
+        quantity: z
+            .number({
+                invalid_type_error: t(`${translationPrefix}.upsert.form.items.quantity.errors.required`),
+            })
+            .min(0.001, t(`${translationPrefix}.upsert.form.items.quantity.errors.min`))
+            .refine((val) => !isNaN(val), {
+                message: t(`${translationPrefix}.upsert.form.items.quantity.errors.invalid`),
+            }),
+        unitPrice: z
+            .number({
+                invalid_type_error: t(`${translationPrefix}.upsert.form.items.unitPrice.errors.required`),
+            })
+            .min(0, t(`${translationPrefix}.upsert.form.items.unitPrice.errors.min`))
+            .refine((val) => !isNaN(val), {
+                message: t(`${translationPrefix}.upsert.form.items.unitPrice.errors.invalid`),
+            }),
+        vatRate: z
+            .number({
+                invalid_type_error: t(`${translationPrefix}.upsert.form.items.vatRate.errors.required`),
+            })
+            .min(0, t(`${translationPrefix}.upsert.form.items.vatRate.errors.min`)),
+        order: z.number(),
+    })
+}
 
 export function InvoiceUpsert({ invoice, open, onOpenChange }: InvoiceUpsertDialogProps) {
     const { t } = useTranslation()
@@ -66,38 +97,7 @@ export function InvoiceUpsert({ invoice, open, onOpenChange }: InvoiceUpsertDial
             .min(0, t("invoices.upsert.form.discountRate.errors.min"))
             .max(100, t("invoices.upsert.form.discountRate.errors.max")),
         items: z.array(
-            z.object({
-                id: z.string().optional(),
-                description: z
-                    .string()
-                    .min(1, t("invoices.upsert.form.items.description.errors.required"))
-                    .refine((val) => val !== "", {
-                        message: t("invoices.upsert.form.items.description.errors.required"),
-                    }),
-                quantity: z
-                    .number({
-                        invalid_type_error: t("invoices.upsert.form.items.quantity.errors.required"),
-                    })
-                    .min(0.001, t("invoices.upsert.form.items.quantity.errors.min"))
-                    .refine((val) => !isNaN(val), {
-                        message: t("invoices.upsert.form.items.quantity.errors.invalid"),
-                    }),
-                unitPrice: z
-                    .number({
-                        invalid_type_error: t("invoices.upsert.form.items.unitPrice.errors.required"),
-                    })
-                    .min(0, t("invoices.upsert.form.items.unitPrice.errors.min"))
-                    .refine((val) => !isNaN(val), {
-                        message: t("invoices.upsert.form.items.unitPrice.errors.invalid"),
-                    }),
-                vatRate: z
-                    .number({
-                        invalid_type_error: t("invoices.upsert.form.items.vatRate.errors.required"),
-                    })
-                    .min(0, t("invoices.upsert.form.items.vatRate.errors.min")),
-                type: z.enum(['HOUR', 'DAY', 'DEPOSIT', 'SERVICE', 'PRODUCT']).optional(),
-                order: z.number(),
-            }),
+            createItemSchema(t, "invoices", z.enum(['HOUR', 'DAY', 'DEPOSIT', 'SERVICE', 'PRODUCT']).optional()),
         ),
     })
 
@@ -121,38 +121,7 @@ export function InvoiceUpsert({ invoice, open, onOpenChange }: InvoiceUpsertDial
         currency: z.string().optional(),
         autoSend: z.boolean().optional(),
         items: z.array(
-            z.object({
-                id: z.string().optional(),
-                description: z
-                    .string()
-                    .min(1, t("recurringInvoices.upsert.form.items.description.errors.required"))
-                    .refine((val) => val !== "", {
-                        message: t("recurringInvoices.upsert.form.items.description.errors.required"),
-                    }),
-                type: z.string(),
-                quantity: z
-                    .number({
-                        invalid_type_error: t("recurringInvoices.upsert.form.items.quantity.errors.required"),
-                    })
-                    .min(0.001, t("recurringInvoices.upsert.form.items.quantity.errors.min"))
-                    .refine((val) => !isNaN(val), {
-                        message: t("recurringInvoices.upsert.form.items.quantity.errors.invalid"),
-                    }),
-                unitPrice: z
-                    .number({
-                        invalid_type_error: t("recurringInvoices.upsert.form.items.unitPrice.errors.required"),
-                    })
-                    .min(0, t("recurringInvoices.upsert.form.items.unitPrice.errors.min"))
-                    .refine((val) => !isNaN(val), {
-                        message: t("recurringInvoices.upsert.form.items.unitPrice.errors.invalid"),
-                    }),
-                vatRate: z
-                    .number({
-                        invalid_type_error: t("recurringInvoices.upsert.form.items.vatRate.errors.required"),
-                    })
-                    .min(0, t("recurringInvoices.upsert.form.items.vatRate.errors.min")),
-                order: z.number(),
-            }),
+            createItemSchema(t, "recurringInvoices", z.string()),
         ),
     })
 
@@ -238,65 +207,8 @@ export function InvoiceUpsert({ invoice, open, onOpenChange }: InvoiceUpsertDial
         }
     }, [invoice, form, recurringForm, isEdit])
 
-    const { control, handleSubmit, setValue } = form
-    const { fields, append, move, remove } = useFieldArray({
-        control,
-        name: "items",
-    })
-
-    const { control: recurringControl, handleSubmit: handleRecurringSubmit, setValue: setRecurringValue } = recurringForm
-    const { fields: recurringFields, append: appendRecurring, move: moveRecurring, remove: removeRecurring } = useFieldArray({
-        control: recurringControl,
-        name: "items",
-    })
-
-    const sensors = useSensors(useSensor(MouseSensor), useSensor(TouchSensor))
-
-    const onDragEnd = (event: any) => {
-        const { active, over } = event
-        if (active.id !== over?.id) {
-            const oldIndex = fields.findIndex((f) => f.id === active.id)
-            const newIndex = fields.findIndex((f) => f.id === over.id)
-            move(oldIndex, newIndex)
-            const reordered = arrayMove(fields, oldIndex, newIndex)
-            reordered.forEach((_, index) => {
-                setValue(`items.${index}.order`, index)
-            })
-        }
-    }
-
-    const onRecurringDragEnd = (event: any) => {
-        const { active, over } = event
-        if (active.id !== over?.id) {
-            const oldIndex = recurringFields.findIndex((f) => f.id === active.id)
-            const newIndex = recurringFields.findIndex((f) => f.id === over.id)
-            moveRecurring(oldIndex, newIndex)
-            const reordered = arrayMove(recurringFields, oldIndex, newIndex)
-            reordered.forEach((_, index) => {
-                setRecurringValue(`items.${index}.order`, index)
-            })
-        }
-    }
-
-    useEffect(() => {
-        fields.forEach((_, i) => {
-            setValue(`items.${i}.order`, i)
-        })
-    }, [fields, setValue])
-
-    useEffect(() => {
-        recurringFields.forEach((_, i) => {
-            setRecurringValue(`items.${i}.order`, i)
-        })
-    }, [recurringFields, setRecurringValue])
-
-    const onRemove = (index: number) => {
-        remove(index)
-    }
-
-    const onRemoveRecurring = (index: number) => {
-        removeRecurring(index)
-    }
+    const { control, handleSubmit } = form
+    const { control: recurringControl, handleSubmit: handleRecurringSubmit } = recurringForm
 
     const onSubmit = (data: z.infer<typeof invoiceSchema>) => {
         const trigger = isEdit ? updateTrigger : createTrigger
@@ -357,7 +269,7 @@ export function InvoiceUpsert({ invoice, open, onOpenChange }: InvoiceUpsertDial
 
                     <div className="flex-1 overflow-y-auto px-6 py-4">
                         {mode === "invoice" ? (
-                            <Form {...form}>
+                            <Form {...form} key="invoice-form-mode">
                                 <form id="invoice-form" onSubmit={handleSubmit(onSubmit)} className="space-y-4" data-cy="invoice-form">
                                     <FormField
                                         control={control}
@@ -401,33 +313,12 @@ export function InvoiceUpsert({ invoice, open, onOpenChange }: InvoiceUpsertDial
                                         )}
                                     />
 
-                                    <FormField
-                                        control={control}
-                                        name="clientId"
-                                        render={({ field }) => (
-                                            <FormItem>
-                                                <FormLabel required>{t("invoices.upsert.form.client.label")}</FormLabel>
-                                                <FormControl>
-                                                    <SearchSelect
-                                                        options={(clients || []).map((c) => ({ label: c.name || c.contactFirstname + " " + c.contactLastname, value: c.id }))}
-                                                        value={field.value ?? ""}
-                                                        onValueChange={(val) => field.onChange(val || null)}
-                                                        onSearchChange={setClientsSearchTerm}
-                                                        placeholder={t("invoices.upsert.form.client.placeholder")}
-                                                        data-cy="invoice-client-select"
-                                                        noResultsComponent={
-                                                            <Button
-                                                                variant="link"
-                                                                onClick={() => setClientDialogOpen(true)}
-                                                            >
-                                                                {t("invoices.upsert.form.client.noOptions")}
-                                                            </Button>
-                                                        }
-                                                    />
-                                                </FormControl>
-                                                <FormMessage />
-                                            </FormItem>
-                                        )}
+                                    <ClientSelectField
+                                        translationPrefix="invoices"
+                                        dataCy="invoice-client-select"
+                                        clients={clients || []}
+                                        onSearchChange={setClientsSearchTerm}
+                                        onRequestCreateClient={() => setClientDialogOpen(true)}
                                     />
 
                                     <FormField
@@ -540,171 +431,11 @@ export function InvoiceUpsert({ invoice, open, onOpenChange }: InvoiceUpsertDial
                                     />
 
 
-                                    <FormItem>
-                                        <FormLabel>{t("invoices.upsert.form.items.label")}</FormLabel>
-                                        <DndContext sensors={sensors} collisionDetection={closestCenter} onDragEnd={onDragEnd}>
-                                            <SortableContext items={fields.map((f) => f.id)} strategy={verticalListSortingStrategy}>
-                                                <div className="space-y-2">
-                                                    {fields.map((fieldItem, index) => (
-                                                        <SortableItem
-                                                            key={fieldItem.id}
-                                                            id={fieldItem.id}
-                                                            dragHandle={<GripVertical className="cursor-grab text-muted-foreground" />}
-                                                        >
-                                                            <div className="flex gap-2 items-center">
-                                                                <FormField
-                                                                    control={control}
-                                                                    name={`items.${index}.description`}
-                                                                    render={({ field }) => (
-                                                                        <FormItem>
-                                                                            <FormControl>
-                                                                                <Input
-                                                                                    {...field}
-                                                                                    placeholder={t(
-                                                                                        `invoices.upsert.form.items.description.placeholder`,
-                                                                                    )}
-                                                                                />
-                                                                            </FormControl>
-                                                                            <FormMessage />
-                                                                        </FormItem>
-                                                                    )}
-                                                                />
-
-                                                                <FormField
-                                                                    control={control}
-                                                                    name={`items.${index}.type`}
-                                                                    render={({ field }) => (
-                                                                        <FormItem>
-                                                                            <FormControl>
-                                                                                <Select value={field.value ?? 'SERVICE'} onValueChange={(val) => field.onChange(val as any)}>
-                                                                                    <SelectTrigger className="w-32 mb-0" aria-label={t("invoices.upsert.form.items.type.label") as string}>
-                                                                                        <SelectValue />
-                                                                                    </SelectTrigger>
-                                                                                    <SelectContent>
-                                                                                        <SelectItem value="HOUR">{t("invoices.upsert.form.items.type.hour")}</SelectItem>
-                                                                                        <SelectItem value="DAY">{t("invoices.upsert.form.items.type.day")}</SelectItem>
-                                                                                        <SelectItem value="DEPOSIT">{t("invoices.upsert.form.items.type.deposit")}</SelectItem>
-                                                                                        <SelectItem value="SERVICE">{t("invoices.upsert.form.items.type.service")}</SelectItem>
-                                                                                        <SelectItem value="PRODUCT">{t("invoices.upsert.form.items.type.product")}</SelectItem>
-                                                                                    </SelectContent>
-                                                                                </Select>
-                                                                            </FormControl>
-                                                                            <FormMessage />
-                                                                        </FormItem>
-                                                                    )}
-                                                                />
-
-                                                                <FormField
-                                                                    control={control}
-                                                                    name={`items.${index}.quantity`}
-                                                                    render={({ field }) => (
-                                                                        <FormItem>
-                                                                            <FormControl>
-                                                                                <BetterInput
-                                                                                    {...field}
-                                                                                    defaultValue={field.value || ""}
-                                                                                    postAdornment={t(`invoices.upsert.form.items.quantity.unit`)}
-                                                                                    type="number"
-                                                                                    step="0.001"
-                                                                                    placeholder={t(
-                                                                                        `invoices.upsert.form.items.quantity.placeholder`,
-                                                                                    )}
-                                                                                    onChange={(e) =>
-                                                                                        field.onChange(e.target.value === "" ? undefined : Number(e.target.value))
-                                                                                    }
-                                                                                />
-                                                                            </FormControl>
-                                                                            <FormMessage />
-                                                                        </FormItem>
-                                                                    )}
-                                                                />
-
-                                                                <FormField
-                                                                    control={control}
-                                                                    name={`items.${index}.unitPrice`}
-                                                                    render={({ field }) => (
-                                                                        <FormItem>
-                                                                            <FormControl>
-                                                                                <BetterInput
-                                                                                    {...field}
-                                                                                    defaultValue={field.value || ""}
-                                                                                    postAdornment="$"
-                                                                                    type="number"
-                                                                                    step="0.01"
-                                                                                    placeholder={t(
-                                                                                        `invoices.upsert.form.items.unitPrice.placeholder`,
-                                                                                    )}
-                                                                                    onChange={(e) =>
-                                                                                        field.onChange(e.target.value === "" ? undefined : Number(e.target.value))
-                                                                                    }
-                                                                                />
-                                                                            </FormControl>
-                                                                            <FormMessage />
-                                                                        </FormItem>
-                                                                    )}
-                                                                />
-
-                                                                <FormField
-                                                                    control={control}
-                                                                    name={`items.${index}.vatRate`}
-                                                                    render={({ field }) => (
-                                                                        <FormItem>
-                                                                            <FormControl>
-                                                                                <BetterInput
-                                                                                    {...field}
-                                                                                    defaultValue={field.value || 0}
-                                                                                    postAdornment="%"
-                                                                                    type="number"
-                                                                                    step="0.01"
-                                                                                    placeholder={t(
-                                                                                        `invoices.upsert.form.items.vatRate.placeholder`,
-                                                                                    )}
-                                                                                    onChange={(e) =>
-                                                                                        field.onChange(
-                                                                                            e.target.value === ""
-                                                                                                ? undefined
-                                                                                                : Number.parseFloat(e.target.value.replace(",", ".")),
-                                                                                        )
-                                                                                    }
-                                                                                />
-                                                                            </FormControl>
-                                                                            <FormMessage />
-                                                                        </FormItem>
-                                                                    )}
-                                                                />
-
-                                                                <Button variant={"outline"} onClick={() => onRemove(index)}>
-                                                                    <Trash2 className="h-4 w-4 text-red-700" />
-                                                                </Button>
-                                                            </div>
-                                                        </SortableItem>
-                                                    ))}
-                                                </div>
-                                            </SortableContext>
-                                        </DndContext>
-
-                                        <Button
-                                            type="button"
-                                            variant="outline"
-                                            onClick={() =>
-                                                append({
-                                                    description: "",
-                                                    quantity: Number.NaN,
-                                                    unitPrice: Number.NaN,
-                                                    vatRate: Number.NaN,
-                                                    type: 'SERVICE',
-                                                    order: fields.length,
-                                                })
-                                            }
-                                        >
-                                            <Plus className="mr-2 h-4 w-4" />
-                                            {t("invoices.upsert.form.items.addItem")}
-                                        </Button>
-                                    </FormItem>
+                                    <InvoiceLineItemsEditor translationPrefix="invoices" defaultItemType="SERVICE" />
                                 </form>
                             </Form>
                         ) : (
-                            <Form {...recurringForm}>
+                            <Form {...recurringForm} key="recurring-form-mode">
                                 <form id="recurring-invoice-form" onSubmit={handleRecurringSubmit(onSubmitRecurring)} className="space-y-4" data-cy="recurring-invoice-form">
                                     <FormField
                                         control={recurringControl}
@@ -747,33 +478,12 @@ export function InvoiceUpsert({ invoice, open, onOpenChange }: InvoiceUpsertDial
                                         )}
                                     />
 
-                                    <FormField
-                                        control={recurringControl}
-                                        name="clientId"
-                                        render={({ field }) => (
-                                            <FormItem>
-                                                <FormLabel required>{t("recurringInvoices.upsert.form.client.label")}</FormLabel>
-                                                <FormControl>
-                                                    <SearchSelect
-                                                        options={(clients || []).map((c) => ({ label: c.name || c.contactFirstname + " " + c.contactLastname, value: c.id }))}
-                                                        value={field.value ?? ""}
-                                                        onValueChange={(val) => field.onChange(val || null)}
-                                                        onSearchChange={setClientsSearchTerm}
-                                                        placeholder={t("recurringInvoices.upsert.form.client.placeholder")}
-                                                        data-cy="recurring-invoice-client-select"
-                                                        noResultsComponent={
-                                                            <Button
-                                                                variant="link"
-                                                                onClick={() => setClientDialogOpen(true)}
-                                                            >
-                                                                {t("recurringInvoices.upsert.form.client.noOptions")}
-                                                            </Button>
-                                                        }
-                                                    />
-                                                </FormControl>
-                                                <FormMessage />
-                                            </FormItem>
-                                        )}
+                                    <ClientSelectField
+                                        translationPrefix="recurringInvoices"
+                                        dataCy="recurring-invoice-client-select"
+                                        clients={clients || []}
+                                        onSearchChange={setClientsSearchTerm}
+                                        onRequestCreateClient={() => setClientDialogOpen(true)}
                                     />
 
                                     <FormField
@@ -924,167 +634,7 @@ export function InvoiceUpsert({ invoice, open, onOpenChange }: InvoiceUpsertDial
 
                                     <Separator className="my-4" />
 
-                                    <FormItem>
-                                        <FormLabel>{t("recurringInvoices.upsert.form.items.label")}</FormLabel>
-                                        <DndContext sensors={sensors} collisionDetection={closestCenter} onDragEnd={onRecurringDragEnd}>
-                                            <SortableContext items={recurringFields.map((f) => f.id)} strategy={verticalListSortingStrategy}>
-                                                <div className="space-y-2">
-                                                    {recurringFields.map((fieldItem, index) => (
-                                                        <SortableItem
-                                                            key={fieldItem.id}
-                                                            id={fieldItem.id}
-                                                            dragHandle={<GripVertical className="cursor-grab text-muted-foreground" />}
-                                                        >
-                                                            <div className="flex gap-2 items-center">
-                                                                <FormField
-                                                                    control={recurringControl}
-                                                                    name={`items.${index}.description`}
-                                                                    render={({ field }) => (
-                                                                        <FormItem>
-                                                                            <FormControl>
-                                                                                <Input
-                                                                                    {...field}
-                                                                                    placeholder={t(
-                                                                                        "recurringInvoices.upsert.form.items.description.placeholder",
-                                                                                    )}
-                                                                                />
-                                                                            </FormControl>
-                                                                            <FormMessage />
-                                                                        </FormItem>
-                                                                    )}
-                                                                />
-
-                                                                <FormField
-                                                                    control={recurringControl}
-                                                                    name={`items.${index}.type`}
-                                                                    render={({ field }) => (
-                                                                        <FormItem>
-                                                                            <FormControl>
-                                                                                <Select value={field.value ?? 'SERVICE'} onValueChange={(val) => field.onChange(val as any)}>
-                                                                                    <SelectTrigger className="w-32" size="sm" aria-label={t("recurringInvoices.upsert.form.items.type.label") as string}>
-                                                                                        <SelectValue />
-                                                                                    </SelectTrigger>
-                                                                                    <SelectContent>
-                                                                                        <SelectItem value="HOUR">{t("recurringInvoices.upsert.form.items.type.hour")}</SelectItem>
-                                                                                        <SelectItem value="DAY">{t("recurringInvoices.upsert.form.items.type.day")}</SelectItem>
-                                                                                        <SelectItem value="DEPOSIT">{t("recurringInvoices.upsert.form.items.type.deposit")}</SelectItem>
-                                                                                        <SelectItem value="SERVICE">{t("recurringInvoices.upsert.form.items.type.service")}</SelectItem>
-                                                                                        <SelectItem value="PRODUCT">{t("recurringInvoices.upsert.form.items.type.product")}</SelectItem>
-                                                                                    </SelectContent>
-                                                                                </Select>
-                                                                            </FormControl>
-                                                                            <FormMessage />
-                                                                        </FormItem>
-                                                                    )}
-                                                                />
-
-                                                                <FormField
-                                                                    control={recurringControl}
-                                                                    name={`items.${index}.quantity`}
-                                                                    render={({ field }) => (
-                                                                        <FormItem>
-                                                                            <FormControl>
-                                                                                <BetterInput
-                                                                                    {...field}
-                                                                                    defaultValue={field.value || ""}
-                                                                                    postAdornment={t("recurringInvoices.upsert.form.items.quantity.unit")}
-                                                                                    type="number"
-                                                                                    step="0.001"
-                                                                                    placeholder={t(
-                                                                                        "recurringInvoices.upsert.form.items.quantity.placeholder",
-                                                                                    )}
-                                                                                    onChange={(e) =>
-                                                                                        field.onChange(e.target.value === "" ? undefined : Number(e.target.value))
-                                                                                    }
-                                                                                />
-                                                                            </FormControl>
-                                                                            <FormMessage />
-                                                                        </FormItem>
-                                                                    )}
-                                                                />
-
-                                                                <FormField
-                                                                    control={recurringControl}
-                                                                    name={`items.${index}.unitPrice`}
-                                                                    render={({ field }) => (
-                                                                        <FormItem>
-                                                                            <FormControl>
-                                                                                <BetterInput
-                                                                                    {...field}
-                                                                                    defaultValue={field.value || ""}
-                                                                                    postAdornment="$"
-                                                                                    type="number"
-                                                                                    step="0.01"
-                                                                                    placeholder={t(
-                                                                                        "recurringInvoices.upsert.form.items.unitPrice.placeholder",
-                                                                                    )}
-                                                                                    onChange={(e) =>
-                                                                                        field.onChange(e.target.value === "" ? undefined : Number(e.target.value))
-                                                                                    }
-                                                                                />
-                                                                            </FormControl>
-                                                                            <FormMessage />
-                                                                        </FormItem>
-                                                                    )}
-                                                                />
-
-                                                                <FormField
-                                                                    control={recurringControl}
-                                                                    name={`items.${index}.vatRate`}
-                                                                    render={({ field }) => (
-                                                                        <FormItem>
-                                                                            <FormControl>
-                                                                                <BetterInput
-                                                                                    {...field}
-                                                                                    defaultValue={field.value || 0}
-                                                                                    postAdornment="%"
-                                                                                    type="number"
-                                                                                    step="0.01"
-                                                                                    placeholder={t(
-                                                                                        "recurringInvoices.upsert.form.items.vatRate.placeholder",
-                                                                                    )}
-                                                                                    onChange={(e) =>
-                                                                                        field.onChange(
-                                                                                            e.target.value === ""
-                                                                                                ? undefined
-                                                                                                : Number.parseFloat(e.target.value.replace(",", ".")),
-                                                                                        )
-                                                                                    }
-                                                                                />
-                                                                            </FormControl>
-                                                                            <FormMessage />
-                                                                        </FormItem>
-                                                                    )}
-                                                                />
-
-                                                                <Button variant={"outline"} onClick={() => onRemoveRecurring(index)}>
-                                                                    <Trash2 className="h-4 w-4 text-red-700" />
-                                                                </Button>
-                                                            </div>
-                                                        </SortableItem>
-                                                    ))}
-                                                </div>
-                                            </SortableContext>
-                                        </DndContext>
-
-                                        <Button
-                                            type="button"
-                                            variant="outline"
-                                            onClick={() =>
-                                                appendRecurring({
-                                                    description: "",
-                                                    type: "HOUR",
-                                                    quantity: Number.NaN,
-                                                    unitPrice: Number.NaN,
-                                                    vatRate: Number.NaN,
-                                                    order: recurringFields.length,
-                                                })
-                                            }
-                                        >
-                                            <Plus className="mr-2 h-4 w-4" />
-                                            {t("recurringInvoices.upsert.form.items.addItem")}
-                                        </Button>
-                                    </FormItem>
+                                    <InvoiceLineItemsEditor translationPrefix="recurringInvoices" defaultItemType="HOUR" />
 
                                     <Separator className="my-4" />
 
@@ -1135,31 +685,5 @@ export function InvoiceUpsert({ invoice, open, onOpenChange }: InvoiceUpsertDial
                 onCreate={handleClientCreate}
             />
         </>
-    )
-}
-
-function SortableItem({
-    id,
-    children,
-    dragHandle,
-}: {
-    id: string
-    children: React.ReactNode
-    dragHandle: React.ReactNode
-}) {
-    const { attributes, listeners, setNodeRef, transform, transition } = useSortable({ id })
-
-    const style = {
-        transform: CSS.Transform.toString(transform),
-        transition,
-    }
-
-    return (
-        <div ref={setNodeRef} style={style} className="flex items-center gap-2">
-            {children}
-            <div {...attributes} {...listeners}>
-                {dragHandle}
-            </div>
-        </div>
     )
 }
